pyproject.toml
[tool.poetry]
name = "uzc"
version = "0.1.0"
description = "Automation for aligning zoom and Kaltura"
authors = ["misterbianco <jayrad.security@protonmail.com>", "woodenikki <woodenikki@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
PyJWT = "^1.7.1"
KalturaApiClient = "^16.2.0"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
mypy = "^0.770"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
